0.6 Добавил вместо активити фрагменты, это решило проблему с дублированием звука. Внес в Ондестрой остановку медиаплеера, это не выход, но пока сойдет. Приложение уже юзабельное, не падает и ведет себя довольно прилично.
                        еще нужно добавить:
                        1. сортировка mp3 файлов.
                        2. добавить возможность сортировки по альбомам.
                        3.  разобраться с кодировкой
                        4. пофиксить разные баги

+++ app/src/main/java/ua/sinoptik/mediaplayer/AudioList.java
@@ -17,7 +17,7 @@ public class AudioList {
     }
 
     ArrayList<String> duration;
-    ArrayList<Long> id;
+    ArrayList<Integer> id;
     ArrayList<String> title;
     ArrayList<String> artist;
     ArrayList<String> albume;
@@ -61,11 +61,11 @@ public class AudioList {
 
 return id.size();}
 
-    public ArrayList<Long> getId() {
+    public ArrayList<Integer> getId() {
         return id;
     }
 
-    public void setId(ArrayList<Long> id) {
+    public void setId(ArrayList<Integer> id) {
         this.id = id;
     }
 }
